refactor: move fetch stock data method to seperate file
